refactor(packages/eg-walker): split eg-walker internals
