+++ packages/eg-walker/src/engine/indexed-sequence.ts
@@ -1,5 +1,13 @@
-const LEAF_CAPACITY = 64;
-const BRANCH_FACTOR = 32;
+import {
+  BRANCH_FACTOR,
+  LEAF_CAPACITY,
+  createInternal,
+  createLeaf,
+  type IndexedNode,
+  type InternalNode,
+  type ItemLocation,
+  type LeafNode,
+} from "./internals/indexed-sequence-node";
 
 /**
  * Sentinel thrown by the ranked B-tree's prepare/effect index lookups when
@@ -15,84 +23,6 @@ export class IndexOutOfRangeError extends Error {
     this.name = "IndexOutOfRangeError";
   }
 }
-
-type IndexedNode<T extends object> = LeafNode<T> | InternalNode<T>;
-
-interface NodeBase<T extends object> {
-  parent: InternalNode<T> | null;
-  // Cached position within `parent.children`. Kept in sync by every code
-  // path that mutates a parent's child list so {@link positionOfNode}
-  // can walk up without scanning siblings linearly. Meaningless when
-  // `parent` is `null`; we leave it at 0 in that case.
-  childIndex: number;
-  size: number;
-  prepareSum: number;
-  effectSum: number;
-}
-
-interface LeafNode<T extends object> extends NodeBase<T> {
-  readonly kind: "leaf";
-  readonly items: T[];
-  readonly prepareWeights: number[];
-  readonly effectWeights: number[];
-}
-
-interface InternalNode<T extends object> extends NodeBase<T> {
-  readonly kind: "internal";
-  readonly children: IndexedNode<T>[];
-}
-
-interface ItemLocation<T extends object> {
-  leaf: LeafNode<T>;
-  // Cached offset of the item within `leaf.items`. Maintained by every
-  // splice on `leaf.items` so {@link IndexedSequence.positionOf} and
-  // {@link IndexedSequence.updateItem} can skip the O(leaf capacity)
-  // `Array.indexOf` scan.
-  offsetInLeaf: number;
-}
-
-const createLeaf = <T extends object>(): LeafNode<T> => ({
-  kind: "leaf",
-  parent: null,
-  childIndex: 0,
-  items: [],
-  prepareWeights: [],
-  effectWeights: [],
-  size: 0,
-  prepareSum: 0,
-  effectSum: 0,
-});
-
-const createInternal = <T extends object>(
-  children: IndexedNode<T>[],
-): InternalNode<T> => {
-  let size = 0;
-  let prepareSum = 0;
-  let effectSum = 0;
-  for (const child of children) {
-    size += child.size;
-    prepareSum += child.prepareSum;
-    effectSum += child.effectSum;
-  }
-  const node: InternalNode<T> = {
-    kind: "internal",
-    parent: null,
-    childIndex: 0,
-    children,
-    size,
-    prepareSum,
-    effectSum,
-  };
-  for (let index = 0; index < children.length; index++) {
-    const child = children[index];
-    if (!child) {
-      continue;
-    }
-    child.parent = node;
-    child.childIndex = index;
-  }
-  return node;
-};
 
 /**
  * Ranked B-tree for Eg-walker's mutable CRDT sequence.
@@ -470,13 +400,14 @@ export class IndexedSequence<T extends object> {
     let node = this.root;
     let remaining = index;
     while (node.kind === "internal") {
-      const child = node.children.find((candidate) => {
+      let child: IndexedNode<T> | undefined;
+      for (const candidate of node.children) {
         if (remaining < candidate.size) {
-          return true;
+          child = candidate;
+          break;
         }
         remaining -= candidate.size;
-        return false;
-      });
+      }
       if (!child) {
         throw new Error(`Index ${index} out of bounds`);
       }
@@ -532,14 +463,15 @@ export class IndexedSequence<T extends object> {
     let total = 0;
 
     while (node.kind === "internal") {
-      const child = node.children.find((candidate) => {
+      let child: IndexedNode<T> | undefined;
+      for (const candidate of node.children) {
         if (remaining <= candidate.size) {
-          return true;
+          child = candidate;
+          break;
         }
         remaining -= candidate.size;
         total += this.weightSum(candidate, kind);
-        return false;
-      });
+      }
       if (!child) {
         return total;
       }
@@ -589,15 +521,16 @@ export class IndexedSequence<T extends object> {
     let node: IndexedNode<T> = this.root;
 
     while (node.kind === "internal") {
-      const child = node.children.find((candidate) => {
+      let child: IndexedNode<T> | undefined;
+      for (const candidate of node.children) {
         const childSum = this.weightSum(candidate, kind);
         if (remaining < childSum) {
-          return true;
+          child = candidate;
+          break;
         }
         remaining -= childSum;
         position += candidate.size;
-        return false;
-      });
+      }
       if (!child) {
         // Structural-invariant violation: the prepareSum/effectSum
         // aggregate said `index` was reachable, but no child claimed
